requirements.txt
requests
gtts
google-generativeai
python-dotenv
beautifulsoup4
